# contacts-api/src/main/resources/api-spec.yaml
swagger: "2.0"
info:
  description: "A simple contacts API, where a user can get a quick overview of all contacts resources like person, skills."
  version: "1.0.0"
  title: "Contacts API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "stephaneselim94@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/api"
tags:
  - name: "user"
    description: "Everything about users"
  - name: "contact"
    description: "Access to a person's contact"
  - name: "skill"
    description: "Access to a person's skills"
schemes:
  - "http"
paths:
  /contacts:
    get:
      tags:
        - "contact"
      summary: "Get all contacts"
      description: "Get the contacts of everyone"
      operationId: "getContacts"
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Contact"

    post:
      tags:
        - contact
      summary: Add a new contact
      description: Add a new contact for a user
      operationId: addContact
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: New contact
          required: true
          schema:
            $ref: "#/definitions/Contact"
      responses:
        "201":
          description: Contact added
          schema:
            description: Resource URI
            type: object
        "400":
          description: Invalid contact

  /contacts/{contactId}:
    parameters:
      - name: contactId
        in: path
        description: ID of the contact to return
        required: true
        type: integer
        format: int64
    get:
      tags:
        - contact
      summary: Find contact by id
      description: Find a contact using its id in the database.
      operationId: getContactById
      produces:
        - applicaton/json
      responses:
        200:
          description: Contact retrieved
          schema:
            $ref: "#/definitions/Contact"
        400:
          description: Invalid ID
        404:
          description: Contact not found

  /contacts/{contactId}/skills:
    parameters:
      - name: contactId
        in: path
        description: ID of the contact
        required: true
        type: integer
        format: int64

    get:
      tags:
        - "contact"
      summary: Get all skills of a contact
      description: Get all skills of a certain contact as well as their levels for those skills
      operationId: getSkillsByContactId
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SkillLevel"
    post:
      tags:
        - "contact"
      summary: Add a skill to a contact
      description: Add a skill to the set of skills of a contact and set the level
      operationId: addSkillToContact
      produces:
        - application/json
      parameters:
        - in: body
          name: skill
          description: skill to add to contact
          schema:
            $ref: "#/definitions/SkillLevel"
      responses:
        201:
          description: Added a skill to a contact
          schema:
            description: The new skilled contact
            $ref: "#/definitions/SkilledContact"


  /skills:
    get:
      tags:
        - "skill"
      summary: "Get all skills"
      description: "Get all available skills"
      operationId: "getSkills"
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Skill"

    post:
      tags:
        - skill
      summary: Add a new skill
      description: Add a new skill
      operationId: addSkill
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: New skill
          schema:
            $ref: "#/definitions/Skill"
      responses:
        201:
          description: Skill added
          schema:
            description: Resource URI
            type: object
        400:
          description: Invalid Skill

  /skills/{skillId}:
    parameters:
      - name: skillId
        in: path
        description: name of the skill to return
        required: true
        type: string
    get:
      tags:
        - skill
      summary: Find skill
      description: Find skill by its id, which is the name of the skill
      operationId: findSkillById
      produces:
        - application/json
      responses:
        200:
          description: Skill retrieved
          schema:
            $ref: "#/definitions/Skill"
        400:
          description: Invalid Skill Id
        404:
          description: Skill not found

  /users:
    get:
      tags:
        - user
      description: get all users of the API
      summary: get all users
      operationId: getUsers
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"

  /users/{userId}:
    get:
      tags:
        - user
      description: get a user using his id in the database
      summary: get user by id
      operationId: getUserById
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: the id of the user to retrieve
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: User retrieved
          schema:
            $ref: "#/definitions/User"
        400:
          description: Invalid user Id
        404:
          description: user not found



definitions:
  Contact:
    type: object
    properties:
      firstName:
        type: string
      lastName:
        type: string
      fullName:
        type: string
      address:
        type: string
      email:
        type: string
      phoneNumber:
        type: string
  Skill:
    type: object
    properties:
      skill:
        type: string
  SkillLevel:
    type: object
    properties:
      skill:
        type: string
      level:
        type: string
  SkilledContact:
    type: object
    properties:
      contactId:
        type: integer
        format: int64
      skill:
        type: string
      level:
        type: string
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      password:
        type: string
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"